(PC-5380) : Remove core.beneficiaries
Dispatches the content to core.payments for deposit related things
and core.users for the beneficiaries
