+++ src/pcapi/core/payments/exceptions.py
@@ -1,9 +1,5 @@
 from pcapi.models.api_errors import ApiErrors
 
 
-class NotEligible(Exception):
-    pass
-
-
 class AlreadyActivatedException(ApiErrors):
     pass
